requirements.txt
# Discord Bot
py-cord==2.4.1

# Database
SQLAlchemy==2.0.23
aiosqlite>=0.19.0

# Redis
redis==5.0.1

# Image Generation
Pillow==10.1.0
qrcode==7.4.2

# HTTP Client
aiohttp==3.9.1

# Environment variables
python-dotenv==1.0.0

# Logging
python-json-logger==2.0.7

# Type hints
typing-extensions==4.8.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Code quality
black==23.12.0
isort==5.13.2
flake8==6.1.0
mypy==1.7.1

# Development
ipython==8.17.2
